Add support for ERB blocks inside the configuration file
This is useful for applications that are using environment variables and
[dotenv][1] to load different configuration values without touching the YAML
file.

This patch also expands the documentation of how to configure the gem in a
Rails app and removes the `Singleton` usage from the `Configuration` class to
simplify how the class is tested and to avoid leaking state between the multiple
test examples that assert operations inside a config object.

[1] https://github.com/bkeepers/dotenv